[SC-71] chore : dto 패키지에 request, response 분기
