+++ src/main/java/inu/codin/codin/domain/info/domain/office/dto/request/OfficeMemberCreateUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package inu.codin.codin.domain.info.domain.office.dto;
+package inu.codin.codin.domain.info.domain.office.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
