+++ src/main/java/inu/codin/codin/domain/info/domain/office/entity/OfficeMember.java
@@ -1,7 +1,7 @@
 package inu.codin.codin.domain.info.domain.office.entity;
 
 import inu.codin.codin.common.BaseTimeEntity;
-import inu.codin.codin.domain.info.domain.office.dto.OfficeMemberCreateUpdateRequestDto;
+import inu.codin.codin.domain.info.domain.office.dto.request.OfficeMemberCreateUpdateRequestDto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
