+++ src/main/java/inu/codin/codin/domain/info/domain/office/service/OfficeService.java
@@ -1,14 +1,14 @@
 package inu.codin.codin.domain.info.domain.office.service;
 
-import inu.codin.codin.domain.info.domain.office.dto.*;
+import inu.codin.codin.domain.info.domain.office.dto.request.OfficeMemberCreateUpdateRequestDto;
+import inu.codin.codin.domain.info.domain.office.dto.request.OfficeUpdateRequestDto;
+import inu.codin.codin.domain.info.domain.office.dto.response.OfficeDetailsResponseDto;
 import inu.codin.codin.domain.info.domain.office.entity.Office;
 import inu.codin.codin.common.Department;
 import inu.codin.codin.domain.info.domain.office.entity.OfficeMember;
 import inu.codin.codin.domain.info.repository.InfoRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
