+++ src/main/java/inu/codin/codin/domain/info/domain/professor/controller/ProfessorController.java
@@ -2,9 +2,9 @@ package inu.codin.codin.domain.info.domain.professor.controller;
 
 import inu.codin.codin.common.Department;
 import inu.codin.codin.common.ResponseUtils;
-import inu.codin.codin.domain.info.domain.professor.dto.ProfessorCreateUpdateRequestDto;
-import inu.codin.codin.domain.info.domain.professor.dto.ProfessorListResponseDto;
-import inu.codin.codin.domain.info.domain.professor.dto.ProfessorThumbnailResponseDto;
+import inu.codin.codin.domain.info.domain.professor.dto.request.ProfessorCreateUpdateRequestDto;
+import inu.codin.codin.domain.info.domain.professor.dto.response.ProfessorListResponseDto;
+import inu.codin.codin.domain.info.domain.professor.dto.response.ProfessorThumbnailResponseDto;
 import inu.codin.codin.domain.info.domain.professor.service.ProfessorService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
